+++ browser/workspace/preferences.h
@@ -1,0 +1,12 @@
+/* Copyright (c) 2026 The Brave Authors. All rights reserved.
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this file,
+ * You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+#ifndef BRAVE_BROWSER_WORKSPACE_PREFERENCES_H_
+#define BRAVE_BROWSER_WORKSPACE_PREFERENCES_H_
+
+// Profile preference key — stores a dict keyed by hash of the display name.
+inline constexpr char kWorkspacesMetadataPref[] = "brave.workspaces.metadata";
+
+#endif  // BRAVE_BROWSER_WORKSPACE_PREFERENCES_H_
