+++ browser/workspace/workspace_service_factory.h
@@ -36,9 +36,6 @@ class WorkspaceServiceFactory : public ProfileKeyedServiceFactory {
   void RegisterProfilePrefs(
       user_prefs::PrefRegistrySyncable* registry) override;
   bool ServiceIsCreatedWithBrowserContext() const override;
-
-  // BrowserContextKeyedServiceFactory:
-  bool ServiceIsNULLWhileTesting() const override;
 };
 
 #endif  // BRAVE_BROWSER_WORKSPACE_WORKSPACE_SERVICE_FACTORY_H_
